+++ apps/docs/src/app/core/api-files.ts
@@ -26,6 +26,7 @@ export const API_FILES = {
         'AvatarGroupComponent',
         'AvatarGroupItemDirective',
         'AvatarGroupFocusableAvatarDirective',
+        'AvatarGroupPopoverControlDirective',
         'AvatarGroupOverflowBodyDirective',
         'AvatarGroupOverflowItemDirective',
         'AvatarGroupOverflowButtonDirective',
@@ -233,7 +234,8 @@ export const API_FILES = {
         'QuickViewGroupTitleComponent',
         'QuickViewGroupItemComponent',
         'QuickViewGroupItemLabelComponent',
-        'QuickViewGroupItemContentComponent'
+        'QuickViewGroupItemContentComponent',
+        'QuickViewGroupItemContentElementDirective'
     ],
     scrollSpy: ['ScrollSpyDirective'],
     select: ['SelectComponent', 'OptionComponent'],