+++ apps/docs/src/app/core/component-docs/dialog/examples/component-based/dialog-example.component.ts
@@ -10,7 +10,10 @@ import { DialogRef } from '@fundamental-ngx/core/dialog';
             </fd-dialog-header>
 
             <fd-dialog-body>
-                <ul>
+                <p style="text-align: justify; margin: 0">
+                    {{ dialogRef.data.pinnapleDescription }}
+                </p>
+                <ul style="margin-bottom: 0">
                     <li *ngFor="let fact of dialogRef.data.pineappleFunFacts">
                         {{ fact }}
                     </li>
@@ -21,12 +24,14 @@ import { DialogRef } from '@fundamental-ngx/core/dialog';
                 <fd-button-bar
                         label="Interesting"
                         fdType="emphasized"
+                        [compact]="true"
                         (click)="this.dialogRef.close('Continue')">
                 </fd-button-bar>
                 <fd-button-bar
                         label="Cancel"
                         fdInitialFocus
                         fdType="transparent"
+                        [compact]="true"
                         (click)="this.dialogRef.dismiss('Cancel')">
                 </fd-button-bar>
             </fd-dialog-footer>