+++ e2e/wdio/platform/fixtures/appData/object-list-item-contents.ts
@@ -1,4 +1,1 @@
-export const noBorderAttr = 'ng-reflect-no-border';
-export const noBorderStyle = 'border-style';
-export const altNoBorderStyle = 'border-bottom-style';
 export const navUrl = '/platform/home';