+++ libs/platform/src/lib/components/form/form-generator/dynamic-form-generator-select/dynamic-form-generator-select.component.ts
@@ -1,0 +1,16 @@
+import { Component, ViewEncapsulation } from '@angular/core';
+import { BaseDynamicFormGeneratorControl } from '../base-dynamic-form-generator-control';
+import { dynamicFormFieldProvider, dynamicFormGroupChildProvider } from '../providers/providers';
+
+@Component({
+    selector: 'fdp-dynamic-form-generator-select',
+    templateUrl: './dynamic-form-generator-select.component.html',
+    encapsulation: ViewEncapsulation.None,
+    viewProviders: [dynamicFormFieldProvider, dynamicFormGroupChildProvider]
+})
+export class DynamicFormGeneratorSelectComponent extends BaseDynamicFormGeneratorControl {
+
+    constructor() {
+        super();
+    }
+}